pyproject.toml
[tool.poetry]
name = "agabisfastapitask"
version = "0.1.0"
description = ""
authors = ["aigerim678 <aigerimsabyrkhan393@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
uvicorn = {extras = ["standart"], version = "^0.32.1"}
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.10.2"}
pydantic-settings = "^2.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
